requirements.txt
pytest~=7.1.3
PyGithub~=2.2.0
setuptools~=60.2.0